webserver/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
APScheduler==3.9.1
arrow==0.14.7
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.10.0
blueprint==3.4.2
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
Flask==2.0.3
Flask-RESTful==0.3.9
frozenlist==1.3.3
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.3
google-cloud-logging==3.0.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
humanize==4.0.0
ics==0.7
idna==3.3
isodate==0.6.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.11
markdown-it-py==2.2.0
MarkupSafe==2.1.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.4
packaging==21.3
parso==0.8.3
phonenumbers==8.13.10
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pudb==2022.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF4==1.27.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.9
requests==2.27.1
rich==13.3.4
rsa==4.8
sendgrid==6.9.7
serde==0.8.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
starkbank-ecdsa==2.0.3
TatSu==5.6.1
tqdm==4.65.0
twilio==8.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
urwid==2.1.2
urwid-readline==0.13
Werkzeug==2.0.3
yarl==1.8.2
